hw4 done

+++ src/main/java/com/example/taxiservicespring/controller/dto/CarDto.java
@@ -10,6 +10,6 @@ public class CarDto {
     private String regNumber;
     private long modelId;
     private int categoryId;
-    private int locationId;
+    private long locationId;
     private CarStatus status;
 }
